split out TokensController from AssetsController
