+++ src/ComposableController.test.ts
@@ -1,5 +1,7 @@
 import { stub } from 'sinon';
 import type { Patch } from 'immer';
+import { TokensController } from './assets/TokensController';
+import { CollectiblesController } from './assets/CollectiblesController';
 import { AddressBookController } from './user/AddressBookController';
 import { EnsController } from './third-party/EnsController';
 import { ComposableController } from './ComposableController';
@@ -10,7 +12,6 @@ import {
   RestrictedControllerMessenger,
 } from './ControllerMessenger';
 import { PreferencesController } from './user/PreferencesController';
-import { AssetsController } from './assets/AssetsController';
 import {
   NetworkController,
   NetworksChainId,
@@ -90,7 +91,7 @@ describe('ComposableController', () => {
       const preferencesController = new PreferencesController();
       const networkController = new NetworkController();
       const assetContractController = new AssetsContractController();
-      const assetController = new AssetsController({
+      const collectiblesController = new CollectiblesController({
         onPreferencesStateChange: (listener) =>
           preferencesController.subscribe(listener),
         onNetworkStateChange: (listener) =>
@@ -105,24 +106,33 @@ describe('ComposableController', () => {
           assetContractController,
         ),
       });
+      const tokensController = new TokensController({
+        onPreferencesStateChange: (listener) =>
+          preferencesController.subscribe(listener),
+        onNetworkStateChange: (listener) =>
+          networkController.subscribe(listener),
+      });
       const controller = new ComposableController([
         new AddressBookController(),
-        assetController,
+        collectiblesController,
         assetContractController,
         new EnsController(),
         networkController,
         preferencesController,
+        tokensController,
       ]);
       expect(controller.state).toStrictEqual({
         AddressBookController: { addressBook: {} },
         AssetsContractController: {},
-        AssetsController: {
+        CollectiblesController: {
           allCollectibleContracts: {},
           allCollectibles: {},
-          allTokens: {},
           collectibleContracts: [],
           collectibles: [],
           ignoredCollectibles: [],
+        },
+        TokensController: {
+          allTokens: {},
           ignoredTokens: [],
           suggestedAssets: [],
           tokens: [],
@@ -150,7 +160,7 @@ describe('ComposableController', () => {
       const preferencesController = new PreferencesController();
       const networkController = new NetworkController();
       const assetContractController = new AssetsContractController();
-      const assetController = new AssetsController({
+      const collectiblesController = new CollectiblesController({
         onPreferencesStateChange: (listener) =>
           preferencesController.subscribe(listener),
         onNetworkStateChange: (listener) =>
@@ -165,13 +175,20 @@ describe('ComposableController', () => {
           assetContractController,
         ),
       });
+      const tokensController = new TokensController({
+        onPreferencesStateChange: (listener) =>
+          preferencesController.subscribe(listener),
+        onNetworkStateChange: (listener) =>
+          networkController.subscribe(listener),
+      });
       const controller = new ComposableController([
         new AddressBookController(),
-        assetController,
+        collectiblesController,
         assetContractController,
         new EnsController(),
         networkController,
         preferencesController,
+        tokensController,
       ]);
       expect(controller.flatState).toStrictEqual({
         addressBook: {},
